requirements.txt
azure-cognitiveservices-vision-computervision>=0.9.0
azure-identity>=1.14.0
python-dotenv>=1.0.0
numpy>=1.24.3
pillow>=10.1.0
scikit-learn>=1.3.2
pytest>=7.4.3
pytest-cov>=4.1.0
